pyproject.toml
[project]
name = "daqmin"
version = "0.1.0"
description = "Minimal inspector for NI DAQmx"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nidaqmx>=1.0.2",
    "pyside6>=6.8.3",
    "pyyaml>=6.0.2",
    "qtpy>=2.4.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "cjdk>=0.3.0",
    "glom>=24.11.0",
    "hatchling>=1.28.0",
    "pytest>=8.3.5",
    "pytest-qt>=4.4.0",
    "ty>=0.0.14",
]

[project.scripts]
daqmin = "daqmin.__main__:main"
genattrs = "daqmin.genattrs:main"

[tool.hatch.build.targets.wheel.hooks.custom]
path = "src/daqmin/genattrs.py"
dependencies = [
    "cjdk>=0.3.0",
    "glom>=24.11.0",
    "pyyaml>=6.0.2",
]

[tool.ruff]
line-length = 79

[tool.uv]
cache-keys = [
    {file = "pyproject.toml"},
    {file = "src/daqmin/genattrs.py"},
]
